+++ Source/DotNetWorkQueue.Transport.LiteDb.Tests/Basic/LiteDbJobQueueCreationTests.cs
@@ -95,7 +95,8 @@ namespace DotNetWorkQueue.Transport.LiteDb.Tests.Basic
                 createCommand,
                 deleteCommand,
                 scope,
-                connectionManager);
+                connectionManager,
+                new TableNameHelper(connectionInfo));
         }
     }
 }